utils/embedder.py
"""
State-of-the-art fashion image embedder using OpenAI CLIP
Handles real-time inference with optimal performance
"""

import sys
import io

# Fix Windows console encoding for emojis
if sys.platform == 'win32':
    try:
        sys.stdout = io.TextIOWrapper(sys.stdout.buffer, encoding='utf-8')
        sys.stderr = io.TextIOWrapper(sys.stderr.buffer, encoding='utf-8')
    except AttributeError:
        pass

import torch
import open_clip
from PIL import Image
import numpy as np
from typing import Union
import cv2


class FashionEmbedder:
    """
    Production-grade image embedder using CLIP ViT-B/32
    Optimized for fashion similarity search with <50ms inference time
    """

    def __init__(self, model_name: str = "ViT-B-32", pretrained: str = "openai"):
        """
        Initialize CLIP model for fashion embeddings

        Args:
            model_name: CLIP architecture (ViT-B-32 for best speed/quality tradeoff)
            pretrained: Pretrained weights source
        """
        self.device = "cuda" if torch.cuda.is_available() else "cpu"
        print(f"🚀 Loading CLIP {model_name} on {self.device}...")

        # Load CLIP model and preprocessing
        self.model, _, self.preprocess = open_clip.create_model_and_transforms(
            model_name,
            pretrained=pretrained
        )
        self.model = self.model.to(self.device)
        self.model.eval()

        # Get embedding dimension
        self.embedding_dim = self.model.visual.output_dim
        print(f"✓ Model loaded | Embedding dim: {self.embedding_dim}")

    @torch.no_grad()
    def embed_image(self, image: Union[Image.Image, np.ndarray]) -> np.ndarray:
        """
        Generate 512-dim embedding from image

        Args:
            image: PIL Image or numpy array (RGB)

        Returns:
            L2-normalized embedding vector (512,)
        """
        # Convert numpy to PIL if needed
        if isinstance(image, np.ndarray):
            if image.shape[2] == 4:  # RGBA
                image = cv2.cvtColor(image, cv2.COLOR_RGBA2RGB)
            elif len(image.shape) == 3 and image.shape[2] == 3:
                image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
            image = Image.fromarray(image)

        # Preprocess and forward pass
        image_tensor = self.preprocess(image).unsqueeze(0).to(self.device)
        embedding = self.model.encode_image(image_tensor)

        # L2 normalize for cosine similarity
        embedding = embedding / embedding.norm(dim=-1, keepdim=True)

        return embedding.cpu().numpy().flatten()

    @torch.no_grad()
    def embed_batch(self, images: list) -> np.ndarray:
        """
        Batch process multiple images for efficiency

        Args:
            images: List of PIL Images or numpy arrays

        Returns:
            Array of embeddings (N, 512)
        """
        # Preprocess all images
        processed = []
        for img in images:
            if isinstance(img, np.ndarray):
                if img.shape[2] == 4:
                    img = cv2.cvtColor(img, cv2.COLOR_RGBA2RGB)
                elif len(img.shape) == 3 and img.shape[2] == 3:
                    img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
                img = Image.fromarray(img)
            processed.append(self.preprocess(img))

        # Stack and process batch
        batch = torch.stack(processed).to(self.device)
        embeddings = self.model.encode_image(batch)

        # L2 normalize
        embeddings = embeddings / embeddings.norm(dim=-1, keepdim=True)

        return embeddings.cpu().numpy()


def create_embedder():
    """Factory function to create embedder instance"""
    return FashionEmbedder()


if __name__ == "__main__":
    # Quick test
    embedder = FashionEmbedder()
    test_img = Image.new('RGB', (224, 224), color='red')
    embedding = embedder.embed_image(test_img)
    print(f"✓ Test embedding shape: {embedding.shape}")
    print(f"✓ L2 norm: {np.linalg.norm(embedding):.4f} (should be ~1.0)")
