requirements.txt
streamlit>=1.32.0
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.23.0
faiss-cpu>=1.7.4
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
tqdm>=4.65.0
scikit-learn>=1.3.0
